users fixes

+++ backend/app/Http/Controllers/UserController.php
@@ -3,9 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreAvatar;
+use App\Http\Resources\User as UserResource;
 use App\Http\Resources\ViewUser;
 use App\User;
+use Auth;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -17,10 +21,18 @@ class UserController extends Controller
 
     /**
      * Display a listing of the resource.
+     *
+     * @param Request $request
      */
-    public function index()
+    public function index(Request $request)
     {
-        return \App\Http\Resources\User::collection(User::paginate());
+        return UserResource::collection(
+            User::when(!$request->user() || $request->user()->is_admin === false,
+                function ($query) {
+                    return $query->where('is_deleted', false);
+                })
+                ->paginate()
+        );
     }
 
     /**
@@ -57,10 +69,19 @@ class UserController extends Controller
             'password',
             'about_me',
             'is_admin',
+            'is_deleted',
         ]);
 
         if (isset($fields['is_admin']) && !$request->user()->isAdmin()) {
             unset($fields['is_admin']);
+        }
+
+        if (isset($fields['is_deleted']) && !$request->user()->isAdmin()) {
+            unset($fields['is_deleted']);
+        }
+
+        if (isset($fields['password'])) {
+            $fields['password'] = Hash::make($fields['password']);
         }
 
         $user->update($fields);
@@ -86,11 +107,22 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
+     * @param Request $request
      * @param User $user
+     * @throws Exception
      */
-    public function destroy(User $user)
+    public function destroy(Request $request, User $user)
     {
-        $user->delete();
+        if ($request->user()->is_admin) {
+            $user->delete();
+        } else {
+            $user->update([
+                'is_deleted' => true,
+            ]);
+            Auth::logout();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+        }
 
         return response()->json(null, 204);
     }
